adding mypy for type checking
This adds mypy and isort for checking pystemd

+++ pystemd/dbusc.pxd
@@ -8,13 +8,13 @@
 #
 
 from libc.stdint cimport (
-  int16_t,
-  int32_t,
-  uint8_t,
-  int64_t,
-  uint16_t,
-  uint32_t,
-  uint64_t,
+    int16_t,
+    int32_t,
+    int64_t,
+    uint8_t,
+    uint16_t,
+    uint32_t,
+    uint64_t,
 )
